arachnida/requirements.txt
requests
bs4
yarl
Pillow